requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
scikit-image==0.21.0
matplotlib==3.7.2
scipy==1.11.2
colour-science==0.4.3
Pillow==10.0.0
pytest==7.4.2
pandas==2.0.3
pyyaml==6.0.1
argparse
tqdm==4.66.1
seaborn==0.12.2
plotly==5.16.1
colorama==0.4.6